Unified backend calls
Created common functions to make POST and GET requests to backend. Cleaned up dependencies, beautified repo files.
Fixes #10
