+++ app/build.gradle
@@ -53,6 +53,7 @@ dependencies {
     implementation 'com.google.code.gson:gson:2.8.6'
     implementation 'com.google.android.gms:play-services-maps:17.0.0'
     implementation 'com.google.android.gms:play-services-location:17.1.0'
+    implementation 'ru.gildor.coroutines:kotlin-coroutines-okhttp:1.0'
     coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.1.1'
     testImplementation 'junit:junit:4.12'
     testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.9'
