+++ app/src/main/java/com/yalta/repositories/LoginRepo.kt
@@ -1,9 +1,5 @@
 package com.yalta.repositories
 
-import okhttp3.FormBody
-import okhttp3.OkHttpClient
-import okhttp3.Request
-
 class SuccessfulLogin(val token: String) : SuccessfulResponse<SuccessfulLogin>()
 
 interface LoginRepo {
@@ -12,18 +8,7 @@ interface LoginRepo {
 
 class RealLoginRepo : LoginRepo, RealRepo() {
     override suspend fun login(login: String, password: String): RepoResponse<SuccessfulLogin> {
-        val url = "${baseUrl}/login?username=$login&password=$password"
-
-        val client = OkHttpClient()
-        val request = Request.Builder()
-            .url(url)
-            .post(
-                FormBody.Builder()
-                    .build()
-            )
-            .build()
-
-        val response = client.newCall(request).execute()
+        val response = doLoginRequest(login, password)
 
         return response.getRepoResponse { res ->
             SuccessfulLogin(
