+++ app/src/main/java/com/yalta/repositories/PasswordRepo.kt
@@ -1,11 +1,6 @@
 package com.yalta.repositories
 
-import com.yalta.services.SessionService
 import common.ChangePassword
-import okhttp3.MediaType
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.RequestBody
 
 class PasswordChanged : SuccessfulResponse<PasswordChanged>()
 
@@ -15,18 +10,8 @@ interface PasswordRepo {
 
 class RealPasswordRepo : PasswordRepo, RealRepo() {
     override suspend fun changePassword(newPassword: String): RepoResponse<PasswordChanged> {
-        val url = "${baseUrl}/users/me/password"
-        val client = OkHttpClient()
-        val json = common.Serialization.toJson(ChangePassword(newPassword))
-        val request = Request.Builder()
-            .addHeader("Cookie", SessionService.session?.token!!)
-            .url(url)
-            .post(RequestBody
-                .create(MediaType
-                    .parse("text/plain"), json)
-            )
-            .build()
-        val response = client.newCall(request).execute()
+        val body = common.Serialization.toJson(ChangePassword(newPassword))
+        val response = doPostRequest("users/me/password", body)
 
         return response.getRepoResponse {
             PasswordChanged()
