+++ app/src/main/java/com/yalta/repositories/UserRepo.kt
@@ -1,9 +1,6 @@
 package com.yalta.repositories
 
-import com.yalta.services.SessionService
 import common.User
-import okhttp3.OkHttpClient
-import okhttp3.Request
 
 class GotUser(val user: User) : SuccessfulResponse<GotUser>()
 
@@ -13,13 +10,7 @@ interface UserRepo {
 
 class RealUserRepo : UserRepo, RealRepo() {
     override suspend fun getUser(id: Long): RepoResponse<GotUser> {
-        val url = "${baseUrl}/whoami"
-        val client = OkHttpClient()
-        val request = Request.Builder()
-            .addHeader("Cookie", SessionService.session?.token!!)
-            .url(url)
-            .build()
-        val response = client.newCall(request).execute()
+        val response = doGetRequest("whoami")
 
         return response.getRepoResponse { res ->
             GotUser(common.Serialization.fromJson(res.body()?.string()!!, User::class.java))
